syz-manager: pkg: skip seeds with mismatching requirements in fuzzing mode
When running in the test mode, syz-manager already ignores tests that have
arch requirements mismatching the target arch.
Because the same tests are also used as seeds in the fuzzing mode, skip them
likewise, instead of reporting errors if they contain arch-specific syscalls.

The code and tests for parsing the requirements is moved from pkg/runtest to
pkg/manager.
