+++ pkg/runtest/run_test.go
@@ -589,24 +589,3 @@ func TestParsing(t *testing.T) {
 		}
 	}
 }
-
-func TestRequires(t *testing.T) {
-	{
-		requires := parseRequires([]byte("# requires: manual arch=amd64"))
-		if !checkArch(requires, "amd64") {
-			t.Fatalf("amd64 does not pass check")
-		}
-		if checkArch(requires, "riscv64") {
-			t.Fatalf("riscv64 passes check")
-		}
-	}
-	{
-		requires := parseRequires([]byte("# requires: -arch=arm64 manual -arch=riscv64"))
-		if !checkArch(requires, "amd64") {
-			t.Fatalf("amd64 does not pass check")
-		}
-		if checkArch(requires, "riscv64") {
-			t.Fatalf("riscv64 passes check")
-		}
-	}
-}
